HYD-5169 Update socket-worker to use ES6 + Flow
As socket-worker starts to host more application logic (and hopefully an
offline store), Move it over to use static typing and ES6.

We should also be able to clean up the source map output via this
process. Finally we should build the module using prepublish and sending
that to npm.

  - Add babel.
  - Add flow type.
  - Switch from jshint to eslint.
  - Remove old .travis.yml
  - Convert wildcard and test to ES6 + Flow.
  - Convert createSocket and test to ES6 + Flow.
  - Convert getEventSocketHandler and test to ES6 + Flow.
  - Update building to happen on prepublish.
  - Update building to run babel over code.

Signed-off-by: Joe Grund <joe.grund@intel.com>